requirements.txt
biopython==1.66
Django==1.8.3
django-cors-headers==1.1.0
django-countries==3.3
django-environ==0.4.0
django-extensions==1.6.1
django-filter==0.11.0
django-gm2m==0.4.3
django-jsonfield==0.9.15
django-longerusernameandemail==0.5.7
django-model-utils==2.4
django-mptt==0.7.4
django-ordered-model==0.4.2
django-rest-swagger==0.3.4
djangorestframework==3.3.1
drf-extensions==0.2.8
gunicorn==19.4.1
jira==0.50
Markdown==2.6.2
oauthlib==1.0.0
ordereddict==1.1
Pint==0.6
psycopg2==2.6.1
pyparsing==2.0.6
pytz==2015.6
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
simple-salesforce==0.68.0
six==1.9.0
tlslite==0.4.8
django-guardian==1.4.4
